Move command away from third party
